+++ src/main/java/com/darksci/pardot/api/response/campaign/CampaignQueryResponse.java
@@ -46,6 +46,8 @@ public class CampaignQueryResponse {
         }
 
         /**
+         * Campaigns returned by the API.
+         *
          * @return Campaign results.
          */
         public List<Campaign> getCampaigns() {